validation

+++ back/api/Endpoints/Products/Product/UpdateProduct.cs
@@ -6,6 +6,7 @@ using api.RequestsAndResponses.Product.Update;
 using api.Security.Policies;
 using Ardalis.ApiEndpoints;
 using AutoMapper;
+using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -13,27 +14,37 @@ namespace api.Endpoints.Products.Product;
 
 public class UpdateProduct : EndpointBaseAsync
   .WithRequest<UpdateProductRequest>
-  .WithActionResult<BaseProductResponse>
+  .WithActionResult
 {
   private readonly IMapper _mapper;
   private readonly IProductRepo _productRepo;
   private readonly IProductsCategoriesRepo _pcRepo;
   private readonly ICategoryRepo _productCategoryRepo;
+  private readonly IValidator<UpdateProductDto> _validator;
 
-  public UpdateProduct(IMapper mapper, IProductRepo productRepo, IProductsCategoriesRepo pcRepo, ICategoryRepo productCategoryRepo)
+  public UpdateProduct(
+    IMapper mapper,
+    IProductRepo productRepo,
+    IProductsCategoriesRepo pcRepo,
+    ICategoryRepo productCategoryRepo,
+    IValidator<UpdateProductDto> validator)
   {
     _mapper = mapper;
     _productRepo = productRepo;
     _pcRepo = pcRepo;
     _productCategoryRepo = productCategoryRepo;
+    _validator = validator;
   }
 
   [Authorize(Policy = Policies.ManageProducts)]
   [HttpPatch(Routes.Products.ProductRoot)]
-  public override async Task<ActionResult<BaseProductResponse>> HandleAsync(
+  public override async Task<ActionResult> HandleAsync(
     [FromRoute] UpdateProductRequest request,
     CancellationToken cancellationToken = new CancellationToken())
   {
+    var validationResult = await _validator.ValidateAsync(request.Dto, cancellationToken);
+    if (!validationResult.IsValid) throw new BadRequestException(validationResult.ToString());
+
     var existingProduct = await _productRepo.GetOne(request.ProductId);
     if (existingProduct is null) throw new ProductNotFoundException(request.ProductId);
 
@@ -45,13 +56,16 @@ public class UpdateProduct : EndpointBaseAsync
     existingProduct.DiscountAmount = request.Dto.DiscountAmount ?? existingProduct.DiscountAmount;
     existingProduct.IsDiscounted = request.Dto.IsDiscounted ?? existingProduct.IsDiscounted;
 
-    var updated = await _productRepo.Update(existingProduct);
+    await _productRepo.Update(existingProduct);
+
+    if (request.Dto.CategoryId == null && request.Dto.CategoryId == existingProduct.DeepestCategoryId)
+      return Ok();
 
     var oldCategories = await _pcRepo.GetManyByProductId(existingProduct.Id);
     await _pcRepo.DeleteMany(oldCategories);
-    
+
     var allCategories = await _productCategoryRepo.GetAll();
-    
+
     var currentCategory = allCategories.FirstOrDefault(c => c.Id == request.Dto.CategoryId);
     var path = new List<ProductCategory>() {currentCategory};
 
@@ -66,9 +80,9 @@ public class UpdateProduct : EndpointBaseAsync
         lookForParent = false;
         continue;
       }
-      
+
       path.Add(parent);
-      
+
       currentCategory = parent;
     }
 
@@ -81,6 +95,6 @@ public class UpdateProduct : EndpointBaseAsync
       });
     }
 
-    return Ok(_mapper.Map<BaseProductResponse>(updated));
+    return Ok();
   }
 }
